Aeras: First capabilities for enriched meshes for Aeras
I added the enrichMesh() and getMaximumID(stk::mesh::EntityRank)
methods to the Aeras::SpectralDiscretization class.

The enrichMesh() method fills in new enrichedEdges and
enrichedElements member variables, which are analogous in structure to
the wsElNodeID member variable.

The new node IDs are determined with a formula that needs to know the
maximum node ID and edge ID from the original STK mesh (in order to
avoid conflicting IDs).  That functionality is provided by the
getMaximumID(stk::mesh::EntityRank) method.

All this commit provides is these two methods, which I have gotten to
compile successfully.  The Aeras::SpectralDiscretization::updateMesh()
method does not yet call it.  Future work will be to add this and then
have the other methods called by updateMesh()
(e.g. computeOwnedNodesAndUnknowns(),
computeOverlapNodesAndUnknowns(), transformMesh(), etc) use
enrichedElements in addition to metaData and bulkData to do their
computations and fills.