Implement ScatterResidual<Tangent>.
To do this, I introduced ScalarRefT into PHAL::AlbanyTraits::{Residual,
Jacobian, Tangent, DistParamDeriv}. Their definitions may change, but the
important thing is that PHAL::AlbanyTraits::Residual::ScalarRefT and
PHAL::AlbanyTraits::Jacobian::ScalarRefT, etc., exist. This lets us treat a
reference to a ScalarT in a uniform way. In particular, this pattern can be
used:
    ScalarRefT
      valptr = (this->tensorRank == 0 ? this->val[eq](cell,node) :
                this->tensorRank == 1 ? this->valVec[0](cell,node,eq) :
                this->valTensor[0](cell,node, eq/numDim, eq%numDim));

With this commit, and with Albany built with
    -D ENABLE_KOKKOS:BOOL=OFF \
I get this for SteadyHeat3D:
    Start 1: SteadyHeat3D
1/4 Test #1: SteadyHeat3D .....................   Passed    4.10 sec
    Start 2: SteadyHeat3D_10x10x10_ioss
2/4 Test #2: SteadyHeat3D_10x10x10_ioss .......   Passed    7.70 sec
    Start 3: SteadyHeat3D_10x10x10_ascii
3/4 Test #3: SteadyHeat3D_10x10x10_ascii ......   Passed    8.30 sec
    Start 4: SteadyHeat3D_Tpetra
4/4 Test #4: SteadyHeat3D_Tpetra ..............   Passed    3.83 sec