Modify extruded STK mesh import.
The previous implementation was correct, but in the new-Kokkos build,
MultiVector::getVector(i) seems broken in a certain case (I sent a simple
breaking case, if indeed I'm right, to the Tpetra/Kokkos folks).

It happens that the previous impl does imports on each vector of the
multivector. No reason not to do the import all at once. That works in the
new-Kokkos build, and it's probably a little better anyway, so I'm modifying the
code here and then will merge it into tpetra_kokkos.