Several things related to reference_type and operator[].
  1. In PHAL_AlbanyTraits, introduce PHAL::Ref<ScalarT>::type.

  2. operator[] for runtime MDField is going to be deprecated, and its
implementation is currently in flux. Hence I've introduced two means to loop
over an MDField<ScalarT>. One of these means also works for compile-time
MDField<ScalarT, ...>. To be clear, since layout of the array is not known,
order of iteration cannot be defined, just like with operator[].

  2a. The first is an iterator of sorts,
    template<typename ScalarT> PHAL::MDFieldIterator.
Importantly, its data is all stack-allocated. It's used like this:
    // m is an MDField<ScalarT>
    // a is a std::vector<RealType>, say.
    PHAL::MDFieldIterator<ScalarT> mi(m);
    do {
      *mi = a[mi.idx()];
    } while (++mi);
*mi returns PHAL::Ref<ScalarT>::type, and mi.idx() returns in incremented
integer. The implementation uses operator() rather than operator[]. This
iterator works only on runtime MDField at the moment.

  2b. The second is more efficient and works on compile-time and runtime
MDField. You write a functor having an operator() something like this:
    template<typename ScalarT>
    struct Loader {
      const std::vector<RealType>& v;
      Loader (const std::vector<RealType>& v) : v(v) {}
      void operator() (typename PHAL::Ref<ScalarT>::type a, const int i) {
        a = v[i];
      }
    };
Then you call
    Loader<ScalarT> loader(v);
    PHAL::loop(loader, A);
The idea is that the functor is given references to array entries. I use
PHAL::loop to implement PHAL::scale(A, v) and PHAL::set(A, v), for example.

  PHAL_Utilities is a bit of a mess right now. I'll clean it up once all the
design work settles down.
  reduceAll in PHAL_Utilities.cpp is still in progress.

  GPAM_Tpetra is in better shape than it was. Now all comparisons pass (down
from 8 failures), but there is a seg fault at the end. In any case, the
reduceAll isn't implemented, so it gets the right responses purely by chance.