For the NO_KOKKOS_ALBANY version of ScatterResidual, go back to previous impl.
There seems to be a problem with the implementation change to the
NO_KOKKOS_ALBANY Jacobian specialization that was checked in earlier today. It
may not be an Albany issue, but it will require investigation. In any case, I
would like to keep the NO_KOKKOS_ALBANY code paths working and all ctests
passing on these code paths. The purpose of NO_KOKKOS_ALBANY is to make Albany
work with SerialNode against the new Kokkos, Phalanx, Sacado, and Intrepid
versions in the same way it worked against the old versions; speedup is
secondary for this set of code paths.

For this reason, this commit returns the NO_KOKKOS_ALBANY implementation of the
Jacobian specialization to its original form.