Bug fix in enrichMesh() function in Aeras::SpectralDiscretization.
The loop identifying the enriched nodes was using edgeNodes, which came from STK
and therefore was for the linear mesh that was read in.  Instead, the enrichedNodes array
should be called.  With this fix, enrichment code runs for arbitrary Points Per Edge
value (I have tested up to 15), albeit for now only in serial.  There is a seg fault
in computeOverlapNodesAndUnknowns() occuring for all Points Per Edge values that needs to be debugged.