Implemented residual, viscosity and source term for FELIX FO x-z test case for GMD paper.
Changed problem definition for FELIX FO Stokes s.t. neq (# dofs / node) is set based on the
name "Equation Set" in the input file.  By default, this is "FELIX" which corresponds to the usual
2 dofs/node, but it can also be "FELIX X-Z" or "Poisson", which has 1 dofs / node (the Poisson
physics was a part of early debugging and is now effectively obsolete).

Remaining to-dos for x-z test case:

1.) Add transformMesh() routine.
2.) Add appropriate (Neumann) BCs.
3.) Check messy analytic expressions for typos!