Modify FieldManagerScalarResponseFunction to handle multiple element blocks.
Previously, the response field manager in FieldManagerScalarResponseFunction
evaluated every workset, unlike the field and state field managers in
Application. We want the rfm to be able to evaluate only the worksets belonging
to the rfm's element block. We also want to retain the previous behavior, as
some response functions want access to all element blocks.

This commit provides this new behavior. It is controlled by an option whose
default value (false) preserves the previous behavior:
      <ParameterList name="ResponseParams 0">
        <Parameter name="Restrict to Element Block" type="bool" value="true"/>
      </ParameterList>

Additionally, for IPToNodalField and (later, once I get to it)
ProjectIPToNodalField, it never makes sense to have the parameter value other
than true. Hence in ResponseFactory, I make the default true for
IPToNodalField. (ProjectIPToNodalField is still using the eb-hack code until I
get to reworking it.)