Added explicit node sharing to be compatible with new STK changes
STK mesh has removed the capability to automatically figure out
sharing of node across domains as part of modification_end().
They added a utility function to manually perform this function,
which I added before almost all occurances of modification_end().
This was NOT needed for the exodus/Ioss mesh setup since the
sharing is done explicitly by these utilities.

By the name of this function, it is apparent that a more
permanent solution will be needed.
+  stk::mesh::fix_node_sharing_delete_on_2015_03_06(get_bulk_data());