Towards hooking together Shards CellTopologyData and enriched spectral element meshes for Aeras:
1.) Added an AerasMeshSpecsStruct struct to Aeras_SpectralDiscretization.hpp
that takes in an Albany::MeshSpecsStruct, copies it and will overwrite the object's
cell topology to replace the original one with an enriched one.

2.) Put logic in Albany_DiscretizationFactory.cpp to return meshSpecs from
the meshStruct object only if the method is not "Ioss Aeras" or "Exodus Aeras"
(the spectral discretization methods).  If the method is "Ioss Aeras"
or "Exodus Aeras", the meshSpecs returned are from the new AerasMeshSpecsStruct struct.

What remains to be done is to create the new enriched Shards CellTopology object
in the AerasMeshSpecsStruct createAerasMeshSpecs function.