Aeras: Cleanup of Aeras_SpectralDiscretization.cpp
Got rid of some unneeded variables and code commented out due to
debugging.  I tried this on four processors, and I get a new
segmentation fault, this one different than what I was debugging
before.  STK raises this exception, in Bucket.hpp, complaining of an
index out of bounds.  I will continue to hunt it down, but this
version of the code is better than previous versions.