LCM: Continuing work on coupled Schwarz model evaluator.
Redesigned LCM::SchwarzMultiscale class constructor so it takes in
a parameterlist, comm object, and initial_guess vector, like other
similar classes (e.g., QCAD::CoupledPoissonSchrodinger), rather than
an array of model evaluators.  The array of model evaluators is instead
created in the constructor by reading the "Coupled System" line in
the input file, e.g.,

<ParameterList name="Coupled System">
   <Parameter name="Model XML Files" type="Array(string)" value="{ inputT1.xml, inputT1.xml }"/>
</ParameterList>

An array of apps and models is then created based on the parameter lists in the "Model XML Files"
(inputT1.xml and inputT2.xml in this case).  Code runs and prints relevant correct info about the apps/models
for a test case (to be checked in shortly), but seg faults after the creation of the apps
and models (as expected b/c nothing else is implemented after that yet).

To call the coupled Schwarz model evaluator, the following line should be added to the main input
file:

<Parameter name="Solution Method" type="string" value="Coupled Schwarz"/>

Note that one should run with the AlbanyT executable (and input file based
on the templated software stack), as the solution method logic is not implemented
in the functions used in compiling the Albany executable.

TO DO: put in an error check to prevent user for trying to run this problem with the Albany executable.