Fixing LCM MechanicalCan problem for GPU
There is 1 issue detected:

When we allocate temporary data for the kernel as a MDFIeld <ScalarT, Cell ,Dim , Dim> and then use it inside the kernel we get incorrect results in case Device==CUDA and ScalarT=DFad. This issue needs further investigations.
In purpose to make the kernels produce crrect results I replaced
   MDFIeld<ScalarT, Cell ,Dim , Dim> F
allocation outside of the kernel  with
  ScalarT F[3][3]
inside each kernel