LCM: changed logic in Albany_SolverFactory.cpp for CoupledSchwarz solution method, so that the solver for the coupled model is created and returned.
A new observer class will need to be implemented for the CoupledSchwarz problem.

Commenting out setting of nominal_values in the LCM::CoupledSchwarz constructor
as this'll cause a seg fault now due to some objects not implemented yet and hence not
allocated.  With the modifications to Albany_SolverFactory and this commented out,
we can implement the validation of parameterlists in CoupledSchwarz.cpp .