Aeras: Change data structure for enrichedEdges
The enrichedEdges data member is used in a couple of different places,
and needs to be accessed in different types of loops.  For this to
work, I had to change the data structure from a nested array similar
to wsElNodeID, to a std::map between a GlobalOrdinal (representing the
global edge ID) and a Teuchos::ArrayRCP of GlobalOrdinals
(representing the global node IDs of the nodes on that edge).