Aeras: Fixed bugs in Aeras::SpectralDiscretization
The first bug was that I assumed that the number of edge buckets is
the same as the number of element buckets.  This is not true and to
alter the algorithms for enriching the mesh, I had to move to a new
data structure for enrichedEdges: std::map<GO,Teuchos::ArrayRCP<GO>>.
There are several places in this class where changes had to be made to
accommodate this change.

The second bug was that non-owned elements were being included in
wsElNodeID, which is not correct.  I changed this to include only
locally-owned elements.

This commit has OUTPUT_TO_SCREEN defined.