Aeras: Fix a bug comupting coordinates
The formulas I was using for comupting the coordinates at the node
points was incorrect.  For terms like (x-1.0) and (y+1.0), I had
flipped the addition/subtraction signs, effectively rotating the
coordinates 180 degrees from where they were supposed to be.

This fix seemed to get us over the hump for getting reasonable results
for shallow water.