LCM: continue to work on incorporating hydrogen concentration
As of now we can specify a constant concentration to achieve an increase
in yield stress in the elastoviscoplastic model.

Note that the current implementation of the transport coefficients for the
hydrogen problem produces a circular dependency between that and the
constitutive model. To fix this we will need to break up the computation of
transport coefficients to pre- and post- constitutive model evaluation. I
have not started that yet.