Aeras: adding a variant of the shallow water equations with 1 dof/node (h -- advection-only). This is to facilitate debugging of spectral elements and solve TC1 more efficiently.
To turn on Scalar version of shallow water equations, the following needs to be added to the
parameterlist:

<ParameterList name="Equation Set">
      <Parameter name="Type" type="string" value="Scalar"/>
    </ParameterList>

Please see examples/Aeras/ShallowWater/Sphere10/inputScalarT.xml.

WARNING/FIXME: right now, alpha, a parameter for TC2 is hard-coded to pi/2.
It should be changed such that it's read in from parameterlist at runtime
to allow for a more general problem specification.