Moved deprecated STK node_sharing utility into Albany
History: The capability to do node sharing in STK was removed,
and temporarily replaced with a free function. This
function in STK was marked as deprecated and scheduled for
deletion this week.

In this commit, I moved that function into Albany and called
the new Albany version instead of the temporary version in STK.