Modify comparison test.
This fixes a ctest failure on the RPI cdash.

The previous implementation of scaledCompare returned success if
    fabs(x1 - x2) <= (0.5*(fabs(x1) + fabs(x2)) + fabs(absTol))*relTol.
This effectively made the absolute tolerance absTol*relTol rather than just
absTol. After recent code refactoring, a Felix test fails on one build because
of a quantity that is 0 plus noise at the absolute level of 1e-12. The absolute
tolerance is the default of 1e-8 but is effectively 1e-8*1e-4: hence the
failure. The new implementation of scaledCompare looks like this:
  const double d = fabs(x1 - x2);
  return (d <= 0.5*(fabs(x1) + fabs(x2))*relTol ||
          d <= fabs(absTol)) ? 0 : 1;
It returns success (0) if at least one of the two tests on the relative and
absolute tolerances succeeds, and the test against the absolute tolerance now
uses absTol by itself. I think this new implementation better matches the
intended meanings of the tolerance parameters.