LCM: Modify comparison values in MechanicsTensileTet10 tests.
This and related changes to comparison values are in response to this commit:
    35aaf8e6cbd64d8f4c2af0447a13394269c21d7f
    "LCM: update to the volume averaging procedures for J and pressure"
Relative to this commit:
    aed82d93fdaeb7ff8295556c89185bccebea7db2
    "LCM: loosened tolerances so these tests pass again."
I've opted to tighten the rel tols back to 1e-5 and instead change the
comparison values. My reasoning is that the change Kinematics_Def.hpp results in
new values, but those values should be as stable across builds as the old ones
were. We'll see.