LCM: First pass solution to the eqps dependency, more tests
Right now there are hydrogen transport coefficients that depend on the plastic
strain. For now, we are simply calculating those coefficients based on eqps_old
in order to avoid a dependency issue. We could do better, but I am not sure how
important it will be.

Also, another test was added to verify the conservation of total concentration
as the sum of the lattice concentration and the trapped concentration. Looks
good to 0th order. Other tests were updated, but need more attention.