Adding test for Ortiz-Pandolfi cohesive zone model
This model is employed for adpative insertion because it only focuses
on unloading. The tests does attempt to illustrate unloading from a
finite traction which employs a state variable for linear unloading.
Interpenetration and final failure are also illustrated.

The results are plotted with plot_var.py.

The constitutive model was also changed because I discovered a NaN
through testing (thanks exodiff).