Tightening tolerances and modifying exodiff file
The test was not employing an absolute tolerance. Components of the
Cauchy stress (which are numerically zero) were causing trouble.

I just changed the exodiff file to set a floor for Cauchy stress.

What I was at it, I got a little greedier regarding the residual.
Sorry this was failing. I will check on it the next few days.