Mod/Add in FELIX StokesFO
Added the basal friction coefficient evaluator, with its dedicated parameter
list in the XML file.

Notes:

- now we have to use the new constructBCEvaluators method in BCUtils
  (for the NeumannTraits case)
- I had to add a class HomotopyParamValue, which only stores a static
  pointer to the homotopy parameter. This is used to pass around the
  homotopy parameter to other classes needing it. This is because
  Sacado can only register one parameter for continuation...
- the basal friction coefficien evaluator is built through a method,
  which is templated, so that from the 'constructNeumannEvaluators' method,
  with the help of a private functor, we can build the evaluator for every
  evaluation type, using boost::mpl::for_each and cycling through the
  evaluation types in AlbanyTraits.

We also added a response that saves the field beta, for display purposes.
Notice that, to do this, we had to build the evaluator for beta also in
the 'constructEvaluators' method, so that the evaluator for beta is available
when the field manager arranges the dependencies. If only the BC did not
have a dedicated field manager, we would not need to duplicate this
evaluator...