[RCU] Complete testProjector.
The basic correctness test for the projector is that
    q == interp(M \ b(q)),
where q is a linear function of space and that lives on the integration points,
M is the mass matrix, b(q) is the integral over each element, q* = M \ b(q) is
the L_2 projection onto the nodal points, and interp(q*) is the interpolation
back to the integration points.

This test shows that Projector is implemented correctly. However, the overall
approach does not yet work when Lie transformations are used. That is because I
think I have to deal with phase wrapping in r, where
    F = R S
    r, s = logm(S), logm(S)
and R is a rotation matrix.