LCM: rewrote SchwarzMultiscale class to use Thyra product vectors / spaces, which it turns out is necessary for Belos when a Thyra block matrix is given as the operator defining the problem.  CoupledSchwarz runs on > 1 domain problem :). ||F|| history looks reasonable but still need to check solution for correctness.
Also there is a dyn_cast error from Thyra::DefaultProductVector<double> to
Thyra::TpetraVector<double, int, int, Kokkos::Compat::KokkosDeviceWrapperNode<Kokkos::Serial> >
in the computation of the responses / sensitivities, but this is following the solve (line 163 of Main_SolveT.cpp).
Needs to be resolved.  Likely it's a matter of rewriting the thyra<->tpetra conversion functions in Main_SolveT.cpp.