Nightly build cleanups.
The cdash nightlies died on a link error for two DefaultProductVectorSpace methods.

The cause has two parts. First, two Trilinos _decl.hpp files were included. As a
rule, do not include a F_decl.hpp file; include F.hpp, unless you're doing
something very tricky and you know you want just the decl file.

These days ETI in Trilinos is so good that most or all of us are building with
ETI enabled. That ends up masking the problem of including a _decl.hpp file
because the symbols get defined through ETI. But our nightlies continue to test
the no-ETI build, which I think is a good idea, because that build tends to
catch little linking errors more, as in this case. So that was the second part
of why the nightlies failed: without ETI, the symbols defined in the F_def.hpp
file were never compiled.

I cleaned up other instances of _decl.hpp includes while I was at it. These
snuck in but never caused a problem.