MeshAdapt: Refactor to remove template parameter SizeField.
Previously, AAdapt::MeshAdapt took a template parameter SizeField. I removed it
and made AAdapt::MeshSizeField a pure abstract class to specify what MeshAdapt
expects of a now runtime polymorphic SizeField.

There is one bit of RTTI that I had to use in order to call the right
ma::configure. I speculate that is why SizeField was originally made a template
parameter. But I think this little bit of RTTI is way better than having a
rather awkward template parameter.

Later, I want to do the same for the template parameter Output in
FMDBDiscretization.