Refining input-two-hex test...
Now the test goes through the complete unloading process. Both the elements fully unload and output traction-displacement law for normal separation mirrors the input law.

A python script is included and the resulting .pdf might be used for verification. This could not potentially be converted to a test with Alejandro's approval.

Alejandro, Jake, and I must agree to agree regarding the output of tractions. Right now we output Cauchy stress and 1st P-K traction. Plotting these quantities is problematic in finite deformations.

Per the paper, the Pandolfi-Ortiz model generates 1st P-K tractions because the potential is written in terms of energy per unit reference area.