Addition (FELIX)
The effective pressure N is no longer uniform and user-provided. Instead,
we use the expression N = rho*g*H - p_w, where H is the ice thickness field
(loaded from the mesh) and p_w is the water pressure (from subglacial hydrology).
Moreover, the water pressure is set as a fraction alpha of the ice overburden
pressure rho*g*H. The fraction alpha is given from xml file.

Note: in the next future, p_w will be loaded from file, as a result of the
      solution of the (2D) subglacial hydrology problem.