Aeras: Creating test cases for spectral elements on TC2 now that the spectral elements work [:) :) :)].  These include a test case for 9 node quad spectral elements on the original TC2 "larger" geometry, and test cases on a 24 element mesh obtained using the "Transform Mesh = Cube Sphere" function from a mesh of a cube.  5th order elements are tested on this smaller geometry.
I have tested the elements up to 13th order.  The code finally runs fine in serial and parallel, and
there appears to be spectral convergence with p refinement :).

Caveats:
- To run spectral elements / the spectral element tests, you need to build Albany with ALBANY_EPETRA_EXE turned OFF.
- There is still a bug in the spectral elements when there is > 1 workset.  To get around this,
until it is fixed, one can set "Workset Size" = number of elements in the mesh in Discretization
section of the input file (see examples/Aeras/ShallowWater/TC2/inputSpectralT.xml for an example).