Refine Belos call.
I think there's a missing include in Belos at the moment; that should be fixed
soon. However, I can narrow the scope of what's included in my code, and so
avoid the current problem, so I'm doing that. This should resolve the failed
internal nightly build and prevent a failed RPI nightly build over the weekend.