LCM: Corrected material properties for adaptive insertion. Now both problems, two-hex and two-tet, run properly to completion.
It turns out that by mistake the cohesive energy was massive because
of an unrelistically large maximum opening displacement. This caused
the solver to barf on the two-tet problem but not on the two-hex one.

This change solves that issue for these two basic examples.

Thanks to Andrew Bradley for helping to chase down this one.

More examples to come. More verification needed.