[RCU] Infrastructure to be able to clone and reset an apf::Mesh2.
Thanks to Dan and Brian for tips and Dan for the new apf::disownMdsModel.

The purpose of cloning and resetting is to be able to back up one mesh if
adaptation fails (unlikely) or relaxation/equilibration flips a volume on the
new mesh (a problem).

I forgot to commit separately a change I made to the use of "solution" and
"residual". These were hardcoded, duplicated strings in PUMIMeshStruct and
PUMIDiscretization. I centralized these strings to two static const char* fields
in PUMIMeshStruct. This means a typo will lead to a compile-time error rather
than a run-time one. I also capitalized "Solution" so that it sorts before the
other fields in Paraview, which can save some mouse clicks, e.g. when warping by
vector Solution.