Aeras: Kokkos::vector vs std::vector.
For reasons I'm not going to pin down right now, Kokkos::vector as a replacement
for std::vector is causing problems for the basic serial build all but the
Kokkos performance devs use. This evaluator is the only user of it, it
appears. For now, use the ALBANY_KOKKOS_UNDER_DEVELOPMENT guard in the typical
way to protect it. This gets two Aeras tests working again.