Re-enabled Albany to build without Exodus and netcdf.
Added some "#ifdef ALBANY_SEACAS" protection around code
that requires Trilinos to have been built with Exodus.
This is for a FELIX customer who doesn't want to deal
with netcdf and hdf5 TPLs in Trilinos.

Added logic so that the FELIX tests that don't require
exodus are run when exodus is not compiled into Trilinos.
Previously, we had turned off all FELIX tests, but several
read in ASCII meshes or use STK2D meshes.

I turned off exodus/ioss in Trilinos by removing these lines
as well as the 3 netcdf and 3 hdf5 lines:
     -D Trilinos_ENABLE_SEACASIoss:BOOL=ON \
     -D Trilinos_ENABLE_STKIO:BOOL=ON \
     -D Trilinos_ENABLE_SEACASExodus:BOOL=ON \