MOD: Splitting the function that sets the states arrays in the discretization.
The job is done by a private method that takes as input the StatInfoStruct
and the discretization. This method is then called by the usual interface
method setStateArrays.

The reason is that setStateArrays can now call the new method to set the
states arrays in more than one mesh. For instance (and this is why I did
this), one can set some state arrays in the mesh and some other state
arrays in the boundary mesh (only for Extruded mesh so far).