LCM: Add yield strength dependence on helium
Also re-activated the MechanicsWithHelium test(s), of which there are five.

1) just Mechanics
2) Mechanics and Hydrogen
3) Mechanics and Hydrogen version 2
4) Mecahnics and Helium
5) Mecahnics and Helium version 2

The relevant point is test 4) exercises the Elastoviscoplastic model and the
HeliumODEs, but the coupling coefficient between the helium and plasticity
(alpha2) is set to zero. In test 5), alpha2 is set to a real value, and the
effect of the presence of helium on the plasticity response is demonstrated.