Divided the soon-to-be implicit crystal plasticity integration scheme into smaller logical parts.
The old predictor has been replaced with a function that computes only the change in the slip
increment.  Calculations for the quantities that depend on the slip increment are computed in
separate routines.  This will allow the predictor to be pulled out and replaced with a solve
for the change in slip increment.  Replacing the predictor with a series of smaller functions
also helps with readability and will allow us to, for example, plug-and-play additional hardening
models.

The next step is to construct the matrix containing the derivatives of the (local) residual
vector w.r.t. the slip increments.  This can probably be done with either finite difference
or automatic differentiation.  We'll then be in a position to complete the Newton scheme.