LCM: rework HeliumODes to avoid FPEs
Mostly this was adding logic around all calls to std::cbrt. This issue of course
is that the derivative of sqrt(x) and cbrt(x) is not defined when x = 0, so we
need to do something like the following during the Jacobian computation:

cbrt_x(0.0);
if (x > 0.0) {
  cbrt_x = std::cbrt(x);
}

There is probably a slick way to make this look less awful and not as expensive,
but I am not aware of how to do that.