LCM: Coupling! But there are some issues.
Corrected error in obtaining the solution of the coupled app when
interpolating.

Issue: The solution used in the interpolation is obtained through STK,
and therefore it is the solution of the previous step. Thus the applied
displacements of the true Dirichlet BCs and the displacements of the
Schwarz BCs are out of sync by one step. Need to figure out how to
obtain the most current solution from the solver,
or whoever else has it.

Issue: The system of equations as described by the standard Schwarz
method in Smith, Bjorstad, Gropp, Domain Decomposition, 2004, ch. 1
does not apply directly to Albany and needs to be modified for two
reasons:

1) The domain of the interpolation operator needs to cover the
entire partition in question, including the Dirichlet+Neumann and
Schwarz boundaries. Alternatively, the system needs to be written
in terms of three interpolation operators: one for the interior
of the partition, one for Dirichlet+Neumann and one for Schwarz.

2) In addition, due to the manner in which Dirichlet BCs are
represented in Albany, it tunrs out that the corresponding
linear system described in the above reference cannot be used
as given. Instead, a new system needs to be constructed where
all the Schwarz displacements and interpolation operators are moved
to the RHS. The block diagonal Jacobians continue to be
the identity but the off-diagonal coupling Jacobians simply
become zero. In other words, all the work happens in the Schwarz BCs,
not in the solver. But that means we don't have the most current
solution. See first issue above.