merge: begin PUMI support for large models
this commit is paired with SCOREC/core
commit 6af89bf, which allows an StkModel
to be composed of many geometric entities.

right now the current machinery for StkModels
is still filling them in with one geometric
entity each, but we can soon add a new
input method that does more.

Along the lines of improving performance
in large models, the algorithms in
PUMIDiscretization for element blocks,
sidesets, and nodesets are refactored to
iterate over the mesh only once each
and use geometric classification and upward
adjacency to construct all sets
concurrently.

This in turn requires the BoxBuilder to
construct model topology, using some
newly exposed tools from PUMI
and a surprisingly elegant indicial algorithm.
One .dmg file in the examples was also
filled in with topology.