Initial work on a line search for the crystal plasticity model.
The current line search implementation is definitely not optimal.
But it will hopefully help with debugging.  There is some trickiness
w.r.t. AD.  This initial implementation was written defensively, such
that I'm pretty sure it isn't breaking the AD.  There's room for
improvement.

Also, changed the finite-different tangent (within the CP model)
to be central-difference.  It can be printed to screen along with
the AD tangent for the purpose of comparison.  Or you can run with
the FD tangent, for kicks.