LCM bug fix: Neumann BC construction.
In a multiphysics problem (mechanics + thermo), NeumannBase::calc_press was
using a size of 4, rather than 3 (for 3 dimensions) for
qp_data_returned.dimension(2).

A fix that seems to make sense is to use num_dims_ instead of neq in a few spots
in MechanicsProblem::constructNeumannEvaluators. num_dims_ == neq in the case of
pure mechanics but is > neq in a multiphysics problem.