LCM: Moving to hardening minus recovery and away from power-law hardening
Few reasons:

1. Power-law hardening does not fit material data well
2. The derivative for the power-law model does not play nice with AD

Rather than solve for the ODE for hardening

\dot{kapp} = (H - Rd*kappa)*\dot \gamma

We just solve for it analytically

kappa(gamma) = H/Rd(1 - exp(\gamma*Rd)

and add it to to initial yield tauC

where

g = tauC + kappa

In future work we will honor the ODE and truly make it a state variable

NOTE: The functional form for kappa is identical to another form used in our CP codes

g = tau0 + A*(1 - exp(\gamma*B/A))

where A = H/Rd and B = H.

Why didn't I implement the A/B version?

1. I try to make things as complicated as possible. If you are actually reading this, you are nodding.
2. Linear hardening is not easily accomodated via the A/B approach.

With H/Rd, I just check if Rd is zero and then employ linear hardening. With A/B, I'm checking for Inf.
I want to preserve linear hardening as the default.

More clean up needed regarding input decks and the source. We need to agree on some common terminology.