Aeras: Filled in SpectralDiscretization "Lines" methods
Irina K had stubbed these methods in to support 1D spectral elements,
which in turn support the X-Z model in Aeras.  I filled in almost
everything I think needs to be added, with the exception of
computeWorksetInfoLines(). I still need to check the "Quads" version
of this to determine what needs to be different for the "Lines"
version.