LCM: Added predictor to CrystalPlasticity model...
In an effort to more the integration more robust, a  predictor was added to yield
a better estimate for the slips at n+1 through the slip rate.

An ifdef block was added - SLIP_PREDICTOR

Before Newton, we estimate the slip through

slip_np1 = slip_n + delta t * slip_rate_n

To make this happen the slip rate is now an internal variable.

For the very limited cases examine, it helps a fair amount.

Still much more to do to increase the robustness of the system. There is still
tremendous sensitivity around zero slip and the fact that some slip systems are not activated.

Some more changes in the solver settings - enable more backtracking, etc.

The defalt behavior is that the predictor is ON.