Adding ENSEMBLE capability (formerly "MP") back into Albany.
This capability allows for an ensemble of samples to be propagated
through the finite element assmebly and solves all at once. This
was creatd by Eric Phipps, but needed updating for changes with
the port to Kokkos data structures.

The  ENABLE_SG_MP cmake configuration option no longer exists.
There is now an ENABLE_ENSEMBLE option that turns this new
capability, and a separate ENABLE_SG option that is not
currently working.

There is an ENSEMBLE_SIZE=<int> configuration option (default 32).

There are currently 2 tests of this capabily, only for builds that
are with TriKota in Trilinos and ENABLE_ENSEMBLE=ON in Albany.