Add "ENABLE_BGL" build option ("ON" by default)
If one builds Albany with "-D ENABLE_BGL:BOOL=OFF", this will
disable the use of boost graph inside Albany (along with the
Phalanx evaluator graphs).

The Intel Compiler on the MIC does not compile portions of boost graph, and
it does not appear to be actively maintained currently. This option supports
further progress on the MIC platform without modifying the compiler or boost.