LCM: Handle quadrature for M within ProjectIPtoNodalField.
When the quadrature rule for the problem is not high enough to integrate the
mass matrix, create our own.

This doesn't mean that the projection will be OK. In the case of a quadratic
tet, for example, the rhs, over whose quadrature order ProjectIPtoNodalField has
no control, must be integrated with degree 3 or higher to recover a linear
function exactly. Still, that's one degree lower than the mass matrix.