CMakeLists: Modify cross-compiling procedure.
I've been finding in our various nightlies that the cross-compile configuration
that was introduced a few days ago is messing with things in subtle ways. The
most obvious is that CMAKE_Fortran_IMPLICIT_LINK_LIBRARIES doesn't get set; this
was the direct cause of last night's internal CDash link failure, as Lame is
partly a Fortran library.

I'm introducing a flag ENABLE_CROSS_COMPILE, defaulted to OFF, that implements
the mods from the other day. If it's OFF, then the old (i.e., prior to a few
days ago) build behavior is used.

So those building on the MIC should use the flag -DENABLE_CROSS_COMPILE=ON from
now on. We'll switch back to automagic detection if we figure out a way to do
so.