LCM: Added option to build scripts for specification of compilers through environment variables in addition to their simple self-detection.
STK does not support versions of GCC newer than 4.7.2, which for the most
part does not cause trouble except when compiling with Cuda.

A messy workaround for the time being is to side-install the compiler
that STK wants and inform the build scripts of the existence of that
compiler by means of the standard environment variables CC, CXX and FC.

Thus one can specify the alternate compiler as:

export CXX=/usr/local/gcc-4.7.2/bin/g++

which will be honored by the build scripts instead of using the compilers
that they find.

Use with CAUTION as this change won't prevent conflicting compiler
specifications.

If the variables CC, CXX and FC are unset or set to the native compilers,
the behavior of the build scripts is as before.